Add files via upload
This script allows you to create objects that represent vehicles and then those vehicles are added to a file where they will be stored. That file will represent an available vehicle inventory.

Additionally, we can remove a certain vehicle from that inventory using one of the attributes that is assigned to each object at the time of its creation.